anti-trust refactor thresholds kicked in and core.blockchain was broken out into its own module
