+++ eagleproject/notify/triggers/compound_governoralpha_proposals.py
@@ -6,7 +6,7 @@ from eth_abi import decode_single
 
 from core.common import decode_calls
 from core.addresses import CONTRACT_ADDR_TO_NAME, COMPOUND_GOVERNOR_ALPHA
-from core.sigs import (
+from core.blockchain.sigs import (
     SIG_EVENT_PROPOSAL_CREATED,
     SIG_EVENT_PROPOSAL_CANCELED,
     SIG_EVENT_PROPOSAL_QUEUED,
