+++ eagleproject/notify/triggers/compound_timelock_delay.py
@@ -4,7 +4,7 @@ from eth_utils import decode_hex
 from eth_abi import decode_single
 
 from core.addresses import COMPOUND_TIMELOCK
-from core.sigs import SIG_EVENT_DELAY
+from core.blockchain.sigs import SIG_EVENT_DELAY
 from notify.events import event_high
