+++ eagleproject/notify/triggers/compound_timelock_transactions.py
@@ -5,7 +5,7 @@ from eth_abi import decode_single
 from django.db.models import Q
 
 from core.addresses import CONTRACT_ADDR_TO_NAME, COMPOUND_TIMELOCK
-from core.sigs import (
+from core.blockchain.sigs import (
     SIG_EVENT_CANCEL_TRANSACTION,
     SIG_EVENT_EXECUTE_TRANSACTION,
     SIG_EVENT_QUEUE_TRANSACTION,
