+++ eagleproject/notify/triggers/governable_transfer.py
@@ -3,7 +3,7 @@ from django.db.models import Q
 from eth_utils import decode_hex
 from eth_abi import decode_single
 
-from core.sigs import SIG_EVENT_PENDING_TRANSFER, SIG_EVENT_TRANSFER
+from core.blockchain.sigs import SIG_EVENT_PENDING_TRANSFER, SIG_EVENT_TRANSFER
 from notify.events import event_high
