+++ eagleproject/notify/triggers/proxy_upgraded.py
@@ -3,7 +3,7 @@ from eth_utils import decode_hex
 from eth_abi import decode_single
 
 from core.addresses import CONTRACT_ADDR_TO_NAME
-from core.sigs import SIG_EVENT_UPGRADED
+from core.blockchain.sigs import SIG_EVENT_UPGRADED
 from notify.events import event_high
