+++ eagleproject/notify/triggers/timelock_admin.py
@@ -4,7 +4,10 @@ from eth_utils import decode_hex
 from eth_abi import decode_single
 
 from core.addresses import TIMELOCK
-from core.sigs import SIG_EVENT_NEW_ADMIN, SIG_EVENT_NEW_PENDING_ADMIN
+from core.blockchain.sigs import (
+    SIG_EVENT_NEW_ADMIN,
+    SIG_EVENT_NEW_PENDING_ADMIN,
+)
 from notify.events import event_high
