+++ eagleproject/notify/triggers/vault_buffer.py
@@ -2,7 +2,7 @@
 from decimal import Decimal
 from eth_utils import decode_hex
 from eth_abi import decode_single
-from core.sigs import SIG_EVENT_BUFFER_UPDATE
+from core.blockchain.sigs import SIG_EVENT_BUFFER_UPDATE
 from notify.events import event_normal
