+++ eagleproject/notify/triggers/vault_fees.py
@@ -2,7 +2,7 @@
 from decimal import Decimal
 from eth_utils import decode_hex
 from eth_abi import decode_single
-from core.sigs import SIG_EVENT_REDEEM_FEE
+from core.blockchain.sigs import SIG_EVENT_REDEEM_FEE
 from notify.events import event_normal
