+++ eagleproject/notify/triggers/vault_oracle_changed.py
@@ -1,7 +1,7 @@
 """ Vault fee related events """
 from eth_utils import decode_hex
 from eth_abi import decode_single
-from core.sigs import SIG_EVENT_PRICE_PROVIDER
+from core.blockchain.sigs import SIG_EVENT_PRICE_PROVIDER
 from notify.events import event_normal
