+++ eagleproject/notify/triggers/vault_strategist.py
@@ -2,7 +2,7 @@
 from eth_utils import decode_hex
 from eth_abi import decode_single
 
-from core.sigs import SIG_EVENT_STRATEGIST
+from core.blockchain.sigs import SIG_EVENT_STRATEGIST
 from notify.events import event_normal
